Project/requirements.txt
Django>=3.2,<4.0
djangorestframework
PyMuPDF
python-dotenv
djangorestframework-simplejwt
psycopg2-binary